refactor: reuse code as much as possible
